Fix pedometer service

+++ app/logic/src/main/kotlin/hardcoder/dev/logic/pedometer/PedometerTrackDeleter.kt
@@ -10,6 +10,6 @@ class PedometerTrackDeleter(
 ) {
 
     suspend fun deleteById(id: Int) = withContext(dispatcher) {
-        appDatabase.pedometerTrackQueries.deleteById(id)
+        appDatabase.pedometerTrackQueries.deleteById(id.toLong())
     }
 }
